- Added Blzut3's patch for a real Chex Quest game mode.
SVN r1188 (trunk)

+++ src/win32/i_crash.cpp
@@ -944,7 +944,7 @@ static void AddStackInfo (HANDLE file, void *dumpaddress, DWORD code, CONTEXT *c
 		}
 #endif
 		Writef (file, "  ");
-		for (i = 0; i < max*sizeof(void*); ++i)
+		for (i = 0; i < int(max*sizeof(void*)); ++i)
 		{
 			if (!SafeReadMemory ((BYTE *)scan + i, &peekb, 1))
 			{
